requirements.txt
pyinstaller>=6.0
pywebview>=5.1
